--- common/index.spec.ts
@@ -1,5 +1,0 @@
-describe('ServerlessDi Common', () => {
-  it('Should Build', () => {
-    expect(true).toBeTruthy();
-  });
-});
