--- common/index.ts
@@ -1,5 +1,0 @@
-export { inject, injectable } from 'inversify';
-export * from './src/decorators';
-export * from './src/constants';
-export * from './src/interfaces';
-export * from './src/utils';
