--- common/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "baseUrl": ".",
-    "paths": {
-      "@serverless-di/core": ["../core"],
-      "@serverless-di/common": ["../common"],
-      "@serverless-di/testing": ["../testing"]
-    }
-  }
-}
