+++ core/BUILD
@@ -8,9 +8,12 @@ ts_library(
     srcs = glob(
         [
             "src/**/*.ts",
-            "index.ts",
+            "*.ts",
         ],
         exclude = [
+            '*.spec.ts',
+            '*.test.ts',
+            '*.e2e.ts',
             "src/**/*.test.ts",
             "src/**/*.spec.ts",
             "src/**/*.e2e.ts",
@@ -18,7 +21,6 @@ ts_library(
     ),
     module_name = "@serverless-di/core",
     deps = [
-        "//common:library",
         "@npm//@types/node",
         "@npm//inversify",
         "@npm//reflect-metadata",
@@ -32,7 +34,9 @@ ts_library(
     name = "e2e",
     testonly = True,
     srcs = glob([
-        "index.spec.ts",
+        '*.spec.ts',
+        '*.test.ts',
+        '*.e2e.ts',
         "src/**/*.test.ts",
         "src/**/*.spec.ts",
         "src/**/*.e2e.ts",
