+++ core/index.ts
@@ -1,4 +1,9 @@
+export { inject, injectable } from 'inversify';
 export * from './src/serverless-factory';
-export * from './src/utils/logger.service';
+export * from './src/common/logger.service';
 export * from './src/config';
 export * from './src/aws/dynamoDB.service';
+export * from '../core/src/decorators';
+export * from '../core/src/constants';
+export * from './src/interfaces';
+export * from './src/utils';
