+++ core/src/interfaces/aws.interfaces.ts
