+++ core/src/interfaces/common.interfaces.ts
