+++ core/src/interfaces/index.ts
